moved /events pages inside _authenticated routes tree (folder)
